+++ src/javascript/app_2/Modules/Trading/Components/Form/Purchase/MessageBox/Templates/error_general.jsx
@@ -11,4 +11,4 @@ ErrorGeneral.propTypes = {
     message: PropTypes.string,
 };
 
-export default ErrorGeneral;
+export { ErrorGeneral };
