+++ src/javascript/app_2/Modules/Trading/Components/Form/Purchase/MessageBox/Templates/index.js
@@ -1,0 +1,3 @@
+export * from './error_balance.jsx';
+export * from './error_general.jsx';
+export * from './error_login.jsx';
